Player container w/ item colllector prefab + game Controller prefab, better looting
Player container: Badman has a new child tagged w/ Collector (only one in scene). This is offset to the player's head so that floating loot goes to this point and not the feet.

Game Controller: This has both the respawn mechanic attached, as well as some global variables/scripts required for loot types/ blueprint specs/level item handling. Must use in scene for items to work correctly!

+ Prefab inside Loot folder w/ example Lootable Object.
+ Has destroy script attached which lets player click on the object to destroy it.
+Additionally has a child object that overwrites where the items spawn from (not necessary but preferred).
+ New shake script lets there be some visual feedback when destroying an object.
+ Now the itemController script on Start() will randomize inventory if object or its parent is tagged with Lootable (this used to be in the DestroyObject script).